Can now work with test set and save model. Current architecture gets up to 89% accuracy

+++ src/transform.py
@@ -90,5 +90,5 @@ class ToTensor(object):
         # numpy image: H x W x C
         # torch image: C X H X W
         image = image.transpose((2, 0, 1))
-        return {'image': torch.from_numpy(image),
+        return {'image': torch.from_numpy(image).float(),
                 'generation': gen}
